refactor: rename fabric-manager-monitor to system-service-monitor
Rename the module to reflect its actual scope after the ADR-030
refactoring: it monitors system services and non-DCGM infrastructure
health, not just Fabric Manager.

Renames:
  health-monitors/fabric-manager-monitor/ -> system-service-monitor/
  fabric_manager_monitor (package) -> system_service_monitor
  demos/fabric-manager-monitor/ -> demos/system-service-monitor/
  All K8s manifests, pyproject.toml, Dockerfile, imports updated.

The ADR file (docs/designs/030-fabric-manager-monitor-scope.md) retains
its original name since it documents the historical scope decision.

@deesharma24 -- went with `system-service-monitor` since it covers FM,
GPU services, fabric state, and CUDA context. If you'd prefer
`infrastructure-health-monitor` or another name, happy to rename.
