+++ demos/system-service-monitor/Dockerfile
