+++ demos/system-service-monitor/checks/fabric_check.py
