+++ demos/system-service-monitor/k8s/configmap.yaml
@@ -1,10 +1,10 @@
 apiVersion: v1
 kind: ConfigMap
 metadata:
-  name: fabric-manager-monitor
+  name: system-service-monitor
   namespace: nvsentinel
   labels:
-    app.kubernetes.io/name: fabric-manager-monitor
+    app.kubernetes.io/name: system-service-monitor
     app.kubernetes.io/instance: nvsentinel
     app.kubernetes.io/component: metrics
 data:
@@ -14,11 +14,5 @@ data:
   FLAP_WINDOW: "600"
   FLAP_THRESHOLD: "3"
   ENABLE_FABRIC_CHECK: "true"
-  ENABLE_PCIE_CHECK: "true"
-  ENABLE_CLOCK_CHECK: "true"
-  ENABLE_NVLINK_CHECK: "true"
   ENABLE_CUDA_VALIDATION: "false"
-  CUDA_VALIDATION_INTERVAL: "600"
-  DCGM_EXPORTER_URL: "http://localhost:9400"
-  CLOCK_THROTTLE_RATIO: "0.85"
   LOG_LEVEL: "INFO"
