+++ demos/system-service-monitor/k8s/rbac.yaml
@@ -1,10 +1,10 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: fabric-manager-monitor
+  name: system-service-monitor
   namespace: nvsentinel
   labels:
-    app.kubernetes.io/name: fabric-manager-monitor
+    app.kubernetes.io/name: system-service-monitor
     app.kubernetes.io/instance: nvsentinel
     app.kubernetes.io/component: metrics
 
@@ -12,9 +12,9 @@ metadata:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: fabric-manager-monitor
+  name: system-service-monitor
   labels:
-    app.kubernetes.io/name: fabric-manager-monitor
+    app.kubernetes.io/name: system-service-monitor
     app.kubernetes.io/instance: nvsentinel
     app.kubernetes.io/component: metrics
 rules:
@@ -26,16 +26,16 @@ rules:
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
 metadata:
-  name: fabric-manager-monitor
+  name: system-service-monitor
   labels:
-    app.kubernetes.io/name: fabric-manager-monitor
+    app.kubernetes.io/name: system-service-monitor
     app.kubernetes.io/instance: nvsentinel
     app.kubernetes.io/component: metrics
 roleRef:
   apiGroup: rbac.authorization.k8s.io
   kind: ClusterRole
-  name: fabric-manager-monitor
+  name: system-service-monitor
 subjects:
   - kind: ServiceAccount
-    name: fabric-manager-monitor
+    name: system-service-monitor
     namespace: nvsentinel
