+++ health-monitors/system-service-monitor/system_service_monitor/protos/__init__.py
